Avoid closing and re-opening /dev/urandom [H. Peter Anvin]
There is no point in closing /dev/urandom after using it; we need to
use it multiple times and the cost of holding open a file descriptor
is miniscule at best.

Also, always do the initialization for the last resort fallback using
srand(); if something goes very wrong and we fail to open /dev/urandom
then at least that code is initialized.