Make --dirent work with vfat (issue 14)
From colomb4n@gmail.com:

The patterns used by the --dirent option to overwrite directory names
contained the character " which is not allowed in vfat.

Replace these patterns by similar patterns that are compatible with
vfat.

The patterns used were:

    0x55   01010101    U
    0x22   00100010    "

As I understand this, the idea is to find patterns containing
complementary bits that represent characters allowed in filenames.

Now we use the following patterns which seems to meet the above
requirements while representing characters allowed in vfat:

    0x32   00110010    2
    0x4d   01001101    M