Generate randomness and do I/O in parallel
Initial patch submitted by Richard B. Hill.

Spawn a pthread to generate random data so that I/O and random
generation can be done in parallel.  This is a fairly big performance
win on slow hardware, especially on slower hardware.

If a working pthreads is unavailable, fall back to the old
serial behavior.