unfinished libscrub development
Add tscrub test program for libscrub.  Alter pattern.[ch] interfaces
to work with libscrub.  Implement first cuts of some libscrub functions,
and tweak the scrub.h api slightly.

A large part is still missing: the scrub algorithmic stuff
that is currently in the scrub.c mainline.  This needs to be refactored
before we get very far with the API.

This commit was getting stale and all tests still pass so although
the new functionality is incomplete, I decided to push.