test: fix bashisms in tests
We're using a /bin/sh shebang but the tests aren't
compatible with a POSIX-compliant /bin/sh (like dash,
instead of the usual bash).

Switch to = instead of == which works the same but
works with both Bash and dash.

If desired, we could switch the shebangs to
/bin/bash which would work.

(Notably, Debian defaults to dash as /bin/sh).