Scrub link and remove fix: Scrub and remove both link and target.
This addresses and issue where the -r option removes the
link, but not the target. Now, both will be removed.

In addition, when -D newdir is used it renames the link
but not the target. Now, it will rename the target.

When -D and -r are used together, both link and target
will be removed.